refactor: alert -> notification으로 네이밍 수정
