+++ backend/src/main/java/com/woowacourse/gongcheck/core/application/NotificationService.java
@@ -2,7 +2,7 @@ package com.woowacourse.gongcheck.core.application;
 
 import com.woowacourse.gongcheck.core.application.response.SubmissionCreatedResponse;
 
-public interface AlertService {
+public interface NotificationService {
 
     void sendMessage(final SubmissionCreatedResponse submissionCreatedResponse);
 }
