+++ backend/src/main/java/com/woowacourse/gongcheck/core/application/SubmissionService.java
@@ -34,19 +34,20 @@ public class SubmissionService {
     private final TaskRepository taskRepository;
     private final RunningTaskRepository runningTaskRepository;
     private final SubmissionRepository submissionRepository;
-    private final AlertService alertService;
+    private final NotificationService notificationService;
 
     public SubmissionService(final HostRepository hostRepository, final JobRepository jobRepository,
                              final SpaceRepository spaceRepository, final TaskRepository taskRepository,
                              final RunningTaskRepository runningTaskRepository,
-                             final SubmissionRepository submissionRepository, final AlertService alertService) {
+                             final SubmissionRepository submissionRepository,
+                             final NotificationService notificationService) {
         this.hostRepository = hostRepository;
         this.jobRepository = jobRepository;
         this.spaceRepository = spaceRepository;
         this.taskRepository = taskRepository;
         this.runningTaskRepository = runningTaskRepository;
         this.submissionRepository = submissionRepository;
-        this.alertService = alertService;
+        this.notificationService = notificationService;
     }
 
     @Transactional(isolation = Isolation.SERIALIZABLE)
@@ -56,13 +57,13 @@ public class SubmissionService {
         Job job = jobRepository.getBySpaceHostAndId(host, jobId);
         saveSubmissionAndClearRunningTasks(request, job);
         if (job.hasUrl()) {
-            alertMessage(request, job);
+            sendNotification(request, job);
         }
     }
 
-    private void alertMessage(final SubmissionRequest request, final Job job) {
+    private void sendNotification(final SubmissionRequest request, final Job job) {
         try {
-            alertService.sendMessage(SubmissionCreatedResponse.of(request.getAuthor(), job));
+            notificationService.sendMessage(SubmissionCreatedResponse.of(request.getAuthor(), job));
         } catch (TaskRejectedException e) {
             throw new RuntimeException(e);
         }
